% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigSVM-package.R
\docType{package}
\name{bigSVM-package}
\alias{bigSVM-package}
\title{Solution Paths for Sparse Linear Support Vector Machine with Lasso or
ELastic-Net Regularization}
\description{
Fast algorithm for fitting solution paths of sparse linear SVM with 
lasso or elastic-net regularization to generate sparse solutions and extend 
model fitting with big data that cannot be fully loaded into main memory.
}
\details{
\tabular{ll}{ Package: \tab bigSVM\cr Type: \tab Package\cr Version: \tab
1.0-1\cr Date: \tab 2017-04-07\cr License: \tab GPL-3\cr } Very simple to
use. Accepts \code{X,y} data for binary classification where \code{y}
belongs to +1, -1, and produces the regularized solution path over a grid of
values for the regularization parameter \code{lambda}.
}
\author{
Yaohui Zeng and Congrui Yi
}
\keyword{SVM}
\keyword{classification}
\keyword{learning}
\keyword{machine}
\keyword{models}
